+++ source/Energinet.DataHub.MarketParticipant.Application/Commands/Query/User/UserOverviewFilterDto.cs
@@ -23,4 +23,5 @@ public sealed record UserOverviewFilterDto(
     string? SearchText,
     UserOverviewSortProperty SortProperty,
     SortDirection SortDirection,
+    IEnumerable<Guid> UserRoleIds,
     IEnumerable<UserStatus> UserStatus);