+++ source/Energinet.DataHub.MarketParticipant.Client/source/Energinet.DataHub.MarketParticipant.Client/MarketParticipantClient.cs
@@ -161,6 +161,11 @@ namespace Energinet.DataHub.MarketParticipant.Client
                 userStatus);
         }
 
+        public Task<UserOverviewResultDto> SearchUsersAsync(int pageNumber, int pageSize, UserOverviewFilterDto filter)
+        {
+            return _marketParticipantUserOverviewClient.SearchUsersAsync(pageNumber, pageSize, filter);
+        }
+
         public Task<IEnumerable<SelectionActorDto>> GetSelectionActorsAsync()
         {
             return _marketParticipantActorQueryClient.GetSelectionActorsAsync();