+++ source/Energinet.DataHub.MarketParticipant.Client/source/Energinet.DataHub.MarketParticipant.Client/MarketParticipantUserOverviewClient.cs
@@ -54,5 +54,22 @@ namespace Energinet.DataHub.MarketParticipant.Client
                 .GetJsonAsync<UserOverviewResultDto>()
                 .ConfigureAwait(false);
         }
+
+        public async Task<UserOverviewResultDto> SearchUsersAsync(int pageNumber, int pageSize, UserOverviewFilterDto filter)
+        {
+            var response = await ValidationExceptionHandler
+                .HandleAsync(
+                    () => _clientFactory
+                        .CreateClient()
+                        .Request("useroverview", "users", "search")
+                        .SetQueryParam("pageNumber", pageNumber)
+                        .SetQueryParam("pageSize", pageSize)
+                        .PostJsonAsync(filter))
+                .ConfigureAwait(false);
+
+            return await response
+                .GetJsonAsync<UserOverviewResultDto>()
+                .ConfigureAwait(false);
+        }
     }
 }