+++ source/Energinet.DataHub.MarketParticipant.Domain/Repositories/IUserOverviewRepository.cs
@@ -58,6 +58,7 @@ public interface IUserOverviewRepository
     /// <param name="actorId">The id of the actor.</param>
     /// <param name="searchText">A text to search for, can be empty</param>
     /// <param name="userStatus">Specifies which user status the search should filter on.</param>
+    /// <param name="userRoles">Specifies which user roles the search should filter on.</param>
     /// <returns>A List of users matching the criteria supplied</returns>
     Task<(IEnumerable<UserOverviewItem> Items, int TotalCount)> SearchUsersAsync(
         int pageNumber,
@@ -66,5 +67,6 @@ public interface IUserOverviewRepository
         SortDirection sortDirection,
         Guid? actorId,
         string? searchText,
-        IEnumerable<UserStatus> userStatus);
+        IEnumerable<UserStatus> userStatus,
+        IEnumerable<UserRoleId> userRoles);
 }